pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-flowops"
version = "0.1.0"
description = "AI-first workflow orchestration for commercial intake, risk routing, approvals, and operational handoff."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "pyyaml>=6.0.1",
    "uvicorn[standard]>=0.29.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27.0",
    "pytest>=8.0.0",
    "ruff>=0.5.0",
]

[tool.setuptools]
packages = ["app", "agents", "schemas", "workflows"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["B", "E", "F", "I", "UP"]
